% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natserv-package.R
\docType{data}
\name{nat_states}
\alias{nat_states}
\title{A data.frame with 49 rows and 2 columns}
\description{
\itemize{
  \item state (character) state 2 letter abbreviation
  \item state_name (character) state full name
}
}
\keyword{data}
