% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_data.R
\name{ns_data}
\alias{ns_data}
\title{NatureServe data}
\usage{
ns_data(uid, key = NULL, ...)
}
\arguments{
\item{uid}{(character) a species UID, e.g., ELEMENT_GLOBAL.2.100925}

\item{key}{(character) API key. Required. See \strong{Authentication} below
for more.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a named list, with possible slots:
\itemize{
 \item natureserve_uri
 \item classification
 \item economicAttributes
 \item license
 \item references
 \item conservationStatus
 \item managementSummary
 \item distribution
}
}
\description{
NatureServe data
}
\section{Authentication}{

Get an API key from NatureServe at
\url{https://services.natureserve.org/developer/index.jsp}.
You can pass your token in as an argument or store it one of two places:

\itemize{
  \item your .Rprofile file with an entry like
  \code{options(NatureServeKey = "your-natureserve-key")}
  \item your .Renviron file with an entry like
  \code{NATURE_SERVE_KEY=your-natureserve-key}
}

See \code{\link{Startup}} for information on how to create/find your
.Rrofile and .Renviron files
}
\examples{
\dontrun{
## single id
res <- ns_data(uid = 'ELEMENT_GLOBAL.2.100925')
## many ids at once
res <- ns_data(uid = c('ELEMENT_GLOBAL.2.100925', 'ELEMENT_GLOBAL.2.104470'))
res$ELEMENT_GLOBAL.2.100925
res$ELEMENT_GLOBAL.2.104470
}
}
\references{
\url{https://services.natureserve.org/index.jsp}
}

