% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_ecohier.R
\name{ns_ecohier}
\alias{ns_ecohier}
\title{Get taxon by id}
\usage{
ns_ecohier(uid, ...)
}
\arguments{
\item{uid}{(character) A NatureServe taxon id (The taxon’s Element Global
UID)}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
A list with lots of elements
}
\description{
Get taxon by id
}
\details{
see https://explorer.natureserve.org/api-docs/#_taxon_data_model
for details on the response data
}
\examples{
\dontrun{
ns_ecohier("ELEMENT_GLOBAL.2.683060")
}
}
\references{
https://explorer.natureserve.org/api-docs/
}
