% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_id.R
\name{ns_id}
\alias{ns_id}
\title{Get taxon by uid}
\usage{
ns_id(uid, ...)
}
\arguments{
\item{uid}{(character) A NatureServe taxon id (The taxon’s Element Global
UID). required.}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
A list with lots of elements
}
\description{
Get taxon by uid
}
\details{
see https://explorer.natureserve.org/api-docs/#_taxon_data_model
for details on the response data
}
\examples{
\dontrun{
ns_id("ELEMENT_GLOBAL.2.154701")
}
}
\references{
https://explorer.natureserve.org/api-docs/
}
