% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_map.R
\name{ns_map}
\alias{ns_map}
\alias{ns_map_county}
\alias{ns_map_cons}
\alias{ns_map_stpr}
\title{Map NatureServe data}
\usage{
ns_map_county(x, ...)

ns_map_cons(x, ...)

ns_map_stpr(x, ...)
}
\arguments{
\item{x}{the result of a call to \code{\link[=ns_data]{ns_data()}}}

\item{...}{ignored}
}
\description{
Map NatureServe data
}
\examples{
\dontrun{
# Aquila chrysaetos
x <- ns_data(uid = 'ELEMENT_GLOBAL.2.100925')
ns_map_county(x)
ns_map_cons(x)
ns_map_stpr(x)

# Ursus americanus
x <- ns_data('ELEMENT_GLOBAL.2.100661')
ns_map_county(x)
ns_map_cons(x)
ns_map_stpr(x)
}
}
