% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_search_comb.R
\name{ns_search_comb}
\alias{ns_search_comb}
\title{Combined search}
\usage{
ns_search_comb(
  text = NULL,
  text_adv = NULL,
  status = NULL,
  location = NULL,
  record_type = NULL,
  record_subtype = NULL,
  modified_since = NULL,
  page = NULL,
  per_page = NULL,
  ...
)
}
\arguments{
\item{text}{(character) xxxx}

\item{text_adv}{(list) xxxx}

\item{status}{(character) conservation status, one of G1, G2, G3, G4,
G5, GH, GX, GNR, GNA, GU. case insensitive}

\item{location}{(list) location, country and sub-country}

\item{record_type}{(character) limit results by record type, one of
species or ecosystem}

\item{record_subtype}{(character) limit results by record sub-type}

\item{modified_since}{(character) search for records modified since a
given time. value must be a date and time with a UTC offset in ISO 8601
format. optional}

\item{page}{(integer) Zero-indexed page number; default: 0. optional}

\item{per_page}{(integer) Records per page; default: 20. optional}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Combined search
}
\examples{
\dontrun{
ns_search_comb(text = "robin")
ns_search_comb(text_adv = list(searchToken = "bird",
  matchAgainst = "allNames", operator="similarTo"))
ns_search_comb(status = "G1")
ns_search_comb(location = list(nation = "US"))
ns_search_comb(location = list(nation = "US", subnation = "VA"))
ns_search_comb(record_type = "species")
ns_search_comb(record_subtype = "macrogroup")
ns_search_comb(modified_since = "2020-04-30T00:00:00+0000")
ns_search_comb(page = 0, per_page = 2)
}
}
\references{
https://explorer.natureserve.org/api-docs/
}
\seealso{
Other search: 
\code{\link{ns_search_eco}()},
\code{\link{ns_search_spp}()}
}
\concept{search}
