% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_search_eco.R
\name{ns_search_eco}
\alias{ns_search_eco}
\title{Ecosystem search}
\usage{
ns_search_eco(
  text = NULL,
  text_adv = NULL,
  status = NULL,
  location = NULL,
  ecosystem_taxonomy = NULL,
  record_subtype = NULL,
  modified_since = NULL,
  page = NULL,
  per_page = NULL,
  ...
)
}
\arguments{
\item{text}{(character) xxxx}

\item{text_adv}{(list) xxxx}

\item{status}{(character) conservation status}

\item{location}{(list) location, country and sub-country}

\item{ecosystem_taxonomy}{(character) ecosystem taxonomy}

\item{record_subtype}{(character) limit results by record sub-type}

\item{modified_since}{(character) search for records modified since a
given time. value must be a date and time with a UTC offset in ISO 8601
format. optional}

\item{page}{(integer) Zero-indexed page number; default: 0. optional}

\item{per_page}{(integer) Records per page; default: 20. optional}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Ecosystem search
}
\examples{
\dontrun{
ns_search_eco(text = "robin")
ns_search_eco(text_adv = list(searchToken = "bird", matchAgainst = "allNames", operator="similarTo"))
ns_search_eco(status = "G1")
ns_search_eco(location = list(nation = "US"))
ns_search_eco(location = list(nation = "US", subnation = "VA"))
ns_search_eco(ecosystem_taxonomy = "M067")
ns_search_eco(record_subtype = "macrogroup")
ns_search_eco(modified_since = "2020-04-30T00:00:00+0000")
ns_search_eco(page = 0, per_page = 2)
}
}
\references{
https://explorer.natureserve.org/api-docs/
}
\seealso{
Other search: 
\code{\link{ns_search_comb}()},
\code{\link{ns_search_spp}()}
}
\concept{search}
